+++ xdas/core/routines.py
@@ -932,30 +932,31 @@ def split(da, indices_or_sections="discontinuities", dim="first", tolerance=None
     """
     Split a data array along a dimension.
 
-    Splitting can either be performed at each discontinuity (along interpolated
-    coordinates), at a given set of indices (give as a list of int) or in order to get
-    a given number of equal sized chunks (if a single int is provided).
+    Splitting can either be performed at each discontinuity , at a given set of indices
+    (given as a list of int) or in order to get a given number of equal sized chunks
+    (if a single int is provided).
 
     Parameters
     ----------
     da : DataArray
         The data array to split
-    indices_or_sections : str, int or list of int, optional
-        If `indices_or_section` is an integer N, the array will be divided into N
+    indices_or_sections : str, int or list of int, default="discontinuities"
+        Describe how the splitting must be done:
+        - If `indices_or_section` is an integer N, the array will be divided into N
         almost equal (can differ by one element if the `dim` size is not a multiple of
-        N). If `indices_or_section` is a 1-D array of sorted integers, the entries
+        N).
+        - If `indices_or_section` is a 1-D array of sorted integers, the entries
         indicate where the array is split along `dim`. For example, `[2, 3]` would, for
-        `dim="first"`, result in [da[:2], da[2:3], da[3:]]. If `indices_or_section` is
-        "discontinuities", the `dim` must be an interpolated coordinate and splitting
-        will occurs at locations where they are two consecutive tie_indices with only
-        one index of difference and where the tie_values difference is greater than
-        `tolerance`. Default to "discontinuities".
+        `dim="first"`, result in [da[:2], da[2:3], da[3:]].
+        - If `indices_or_section` is one of "discontinuities", "gaps" or "overlaps",
+        splitting will occurs at the indices given by `Coordinate.get_split_indices`.
     dim : str, optional
         The dimension along which to split, by default "first"
     tolerance : float or timedelta64, optional
-        If `indices_or_sections="discontinuities"` split will only occur on gaps and
-        overlaps that are bigger than `tolerance`. For time coordinates, numeric
-        values are considered as seconds. Zero tolerance by default.
+        Passed to `Coordinate.get_split_indices` if `indices_or_section` is
+        "discontinuities", "gaps" or "overlaps" to determine what can be considered as
+        a discontiuity. For time coordinates, numeric values are considered as seconds.
+        Zero tolerance by default.
 
     Returns
     -------